Limit canvas size by total pixels instead
Rather than clamping each dimension to 32767 pixels, the limit is now at
a total of 1073676289 pixels (32767 squared), but the individual
dimensions can really be anything, allowing for e.g. very tall but
narrow canvases. To avoid totally bonkers sizes, the dimensions are
arbitrarily limited to one million pixels each.

There's been some checks put in place to give proper error messages when
attempting to save or load images that are too large. Newly created
images also get clamped to a reasonable size now to avoid a potential
crash loop stemming from creating a canvas that is too large and then
crashing again when restarting the program.

Relates to #1193.

+++ src/libclient/net/message.cpp
@@ -411,7 +411,7 @@ Message makeSelectionClearMessage(uint8_t contextId, uint8_t selectionId)
 
 Message makeSelectionPutMessage(
 	uint8_t contextId, uint8_t selectionId, uint8_t op, int32_t x, int32_t y,
-	uint16_t w, uint16_t h, const QByteArray &compressedMask)
+	uint32_t w, uint32_t h, const QByteArray &compressedMask)
 {
 	return Message::noinc(DP_msg_selection_put_new(
 		contextId, selectionId, op, x, y, w, h, &Message::setUchars,
